+++ CHANGELOG.md
@@ -2,6 +2,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Add support for drivers with `browser: :remote`.
+
 
 # 2.2.0